requirements.txt
Django==2.0.3
gunicorn==19.7.1
whitenoise==3.3.1
dj-database-url==0.5.0
psycopg2==2.7.4
django-bootstrap4==0.0.6
django-filter==1.1.0
djangorestframework==3.7.7
markdown==2.6.11
celery==4.1.0
flower==0.9.2
django_celery_beat==1.1.1
